Add files via upload
This Terraform folder contains files and codes for deploying an AWS instance (Ubuntu) in multiple regions (free-tier) - specifically in eu-west-1 and eu-central-1. The infrastructure is designed to meet the following requirements:

Deployed across a minimum of 2 availability zones.
Reusable and modularized for multiple environments (dev, staging, prod).
Includes a script that creates Ansible and Docker containers.